chore!: move `src/core/` to `src/simulat/core/`
BREAKING CHANGE: moved the `./src/core` folder to `./src/simulat/core`
