+++ simulat.py
@@ -2,10 +2,10 @@
 # -*- coding: utf-8 -*-
 """Simulat launcher."""
 
-from src.core.game import init
+from src.simulat.core.game import init
 
 if __name__ == "__main__":
     init()
 
-    from src.core.game import simulat
+    from src.simulat.core.game import simulat
     simulat.run()
