+++ src/simulat/core/log_exception.py
