+++ src/simulat/core/surfaces/scenes/game_scene/game_scene.py
@@ -3,9 +3,9 @@
 
 from __future__ import annotations
 
-from src.core.surfaces.scenes.scene import Scene
+from src.simulat.core.surfaces.scenes.scene import Scene
 
-from src.core.surfaces.game_map import GameMap
+from src.simulat.core.surfaces.game_map import GameMap
 
 
 class GameScene(Scene):
