+++ src/simulat/core/time_it.py
